+++ ext/indi-driver/indi-auxremote/auxremotedriver.h
@@ -35,10 +35,10 @@ protected:
   bool Sync(double ra, double dec);
 
   //GUIDE: guiding functions
-  virtual IPState GuideNorth(float ms);
-  virtual IPState GuideSouth(float ms);
-  virtual IPState GuideEast(float ms);
-  virtual IPState GuideWest(float ms);
+  virtual IPState GuideNorth(uint32_t ms);
+  virtual IPState GuideSouth(uint32_t ms);
+  virtual IPState GuideEast(uint32_t ms);
+  virtual IPState GuideWest(uint32_t ms);
 
   bool Park();
   bool UnPark();
